requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.68.0
anyio==4.10.0
asyncer==0.0.8
attrs==25.3.0
audioop-lts==0.2.2
backoff==2.2.1
bidict==0.23.1
certifi==2025.8.3
cffi==2.0.0
chainlit==2.8.0
charset-normalizer==3.4.3
chevron==0.14.0
click==8.3.0
colorama==0.4.6
cryptography==46.0.1
cuid==0.4
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.116.2
filelock==3.19.1
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.9.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
Jinja2==3.1.6
jiter==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-openai==0.3.33
langchain-qdrant==0.2.1
langchain-text-splitters==0.3.11
langsmith==0.4.29
Lazify==0.4.0
literalai==0.1.201
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.14.1
monotonic==1.6
multidict==6.6.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.3.3
openai==1.108.1
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-alephalpha==0.47.2
opentelemetry-instrumentation-anthropic==0.47.2
opentelemetry-instrumentation-bedrock==0.47.2
opentelemetry-instrumentation-chromadb==0.47.2
opentelemetry-instrumentation-cohere==0.47.2
opentelemetry-instrumentation-crewai==0.47.2
opentelemetry-instrumentation-google-generativeai==0.47.2
opentelemetry-instrumentation-groq==0.47.2
opentelemetry-instrumentation-haystack==0.47.2
opentelemetry-instrumentation-lancedb==0.47.2
opentelemetry-instrumentation-langchain==0.47.2
opentelemetry-instrumentation-llamaindex==0.47.2
opentelemetry-instrumentation-logging==0.58b0
opentelemetry-instrumentation-marqo==0.47.2
opentelemetry-instrumentation-mcp==0.47.2
opentelemetry-instrumentation-milvus==0.47.2
opentelemetry-instrumentation-mistralai==0.47.2
opentelemetry-instrumentation-ollama==0.47.2
opentelemetry-instrumentation-openai==0.47.2
opentelemetry-instrumentation-openai-agents==0.47.2
opentelemetry-instrumentation-pinecone==0.47.2
opentelemetry-instrumentation-qdrant==0.47.2
opentelemetry-instrumentation-redis==0.58b0
opentelemetry-instrumentation-replicate==0.47.2
opentelemetry-instrumentation-requests==0.58b0
opentelemetry-instrumentation-sagemaker==0.47.2
opentelemetry-instrumentation-sqlalchemy==0.58b0
opentelemetry-instrumentation-threading==0.58b0
opentelemetry-instrumentation-together==0.47.2
opentelemetry-instrumentation-transformers==0.47.2
opentelemetry-instrumentation-urllib3==0.58b0
opentelemetry-instrumentation-vertexai==0.47.2
opentelemetry-instrumentation-watsonx==0.47.2
opentelemetry-instrumentation-weaviate==0.47.2
opentelemetry-instrumentation-writer==0.47.2
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-semantic-conventions-ai==0.4.13
opentelemetry-util-http==0.58b0
orjson==3.11.3
packaging==25.0
portalocker==3.2.0
posthog==3.25.0
propcache==0.3.2
protobuf==6.32.1
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
pypdf==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-multipart==0.0.20
python-socketio==5.13.0
pywin32==311
PyYAML==6.0.2
qdrant-client==1.15.1
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.27.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sse-starlette==3.0.2
starlette==0.48.0
syncer==2.0.3
tenacity==9.1.2
tiktoken==0.11.0
tokenizers==0.22.1
tomli==2.2.1
tqdm==4.67.1
traceloop-sdk==0.47.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.36.0
watchfiles==0.24.0
wrapt==1.17.3
wsproto==1.2.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.25.0
